+++ apps/decodex/src/agent/tracker_tool_bridge/tests/review_policy/review_policy_handoff_surface/checkpoint_head.rs
@@ -1,0 +1,42 @@
+use crate::agent::tracker_tool_bridge::tests::{
+	self, DynamicToolHandler, FakeLocalRepoInspector, FakePullRequestInspector, FakeTracker,
+	ISSUE_REVIEW_CHECKPOINT_TOOL_NAME, TempDir, TrackerToolBridge, review_policy,
+};
+
+#[test]
+fn review_checkpoint_normalizes_matching_short_head_sha_to_full_head() {
+	let tracker = FakeTracker::new();
+	let issue = tests::sample_issue();
+	let workflow = tests::sample_workflow();
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let review_context = tests::sample_review_context_in(temp_dir.path());
+	let pull_request_inspector = FakePullRequestInspector::new(Vec::new());
+	let local_repo_inspector = FakeLocalRepoInspector::new(vec![Ok(tests::sample_local_repo())]);
+	let bridge = TrackerToolBridge::with_review_handoff_for_test(
+		&tracker,
+		&issue,
+		&workflow,
+		review_context.clone(),
+		&pull_request_inspector,
+		&local_repo_inspector,
+	);
+	let response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_REVIEW_CHECKPOINT_TOOL_NAME,
+		serde_json::json!({
+			"reviewer": "independent_fresh_context",
+			"status": "clean",
+			"head_sha": &tests::sample_local_repo().head_oid[..7],
+			"review_contract": review_policy::handoff_review_contract_json(),
+			"checks": review_policy::review_checks_json(),
+			"evidence": ["Closeout and review policy both point at the current lane head."]
+		}),
+	);
+
+	assert!(response.success);
+	assert!(tracker.comments.borrow().is_empty());
+
+	let checkpoint = tests::persisted_review_policy_checkpoint(&bridge, &issue, &review_context);
+
+	assert_eq!(checkpoint.head_sha(), tests::sample_local_repo().head_oid.as_str());
+}
