+++ apps/decodex/src/agent/tracker_tool_bridge/tests/review_policy/review_policy_handoff_surface/handoff.rs
@@ -1,0 +1,137 @@
+use crate::agent::tracker_tool_bridge::tests::{
+	self, DynamicToolHandler, FakeLocalRepoInspector, FakePullRequestInspector, FakeTracker,
+	ISSUE_REVIEW_HANDOFF_TOOL_NAME, PullRequestDetails, TempDir, TrackerToolBridge,
+};
+
+#[test]
+fn records_review_handoff_and_applies_it_after_validation() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let tracker = FakeTracker::new();
+	let issue = tests::sample_issue();
+	let workflow = tests::sample_workflow();
+	let inspector = FakePullRequestInspector::new(vec![
+		Ok(PullRequestDetails {
+			head_ref_name: String::from("x/decodex-pub-618"),
+			head_ref_oid: String::from("08a20f7dfb9526e7421a5f095b1c6adec84e52d6"),
+			head_repository_name: String::from("decodex"),
+			head_repository_owner: String::from("hack-ink"),
+			is_draft: false,
+			state: String::from("OPEN"),
+			base_ref_name: String::from("main"),
+			url: String::from("https://github.com/hack-ink/decodex/pull/42"),
+		}),
+		Ok(PullRequestDetails {
+			head_ref_name: String::from("x/decodex-pub-618"),
+			head_ref_oid: String::from("08a20f7dfb9526e7421a5f095b1c6adec84e52d6"),
+			head_repository_name: String::from("decodex"),
+			head_repository_owner: String::from("hack-ink"),
+			is_draft: false,
+			state: String::from("OPEN"),
+			base_ref_name: String::from("main"),
+			url: String::from("https://github.com/hack-ink/decodex/pull/42"),
+		}),
+	]);
+	let local_repo_inspector = FakeLocalRepoInspector::new(vec![
+		Ok(tests::sample_local_repo()),
+		Ok(tests::sample_local_repo()),
+	]);
+	let review_context = tests::sample_review_context_in(temp_dir.path());
+	let bridge = TrackerToolBridge::with_review_handoff_for_test(
+		&tracker,
+		&issue,
+		&workflow,
+		review_context.clone(),
+		&inspector,
+		&local_repo_inspector,
+	);
+
+	tests::write_clean_review_checkpoint(&bridge, &issue, &review_context);
+
+	let response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_REVIEW_HANDOFF_TOOL_NAME,
+		serde_json::json!({
+			"pr_url": "https://github.com/hack-ink/decodex/pull/42",
+			"summary": "Implemented the PR-backed review handoff."
+		}),
+	);
+
+	assert!(response.success);
+
+	tests::assert_review_policy_checkpoint_cleared(&bridge, &issue, &review_context);
+
+	bridge.apply_review_handoff().expect("review handoff should apply");
+
+	assert_eq!(tracker.state_updates.borrow().as_slice(), ["state-review"]);
+
+	let comments = tracker.comments.borrow();
+
+	assert_eq!(comments.len(), 1);
+	assert!(comments[0].contains("- pr_url: `https://github.com/hack-ink/decodex/pull/42`"));
+	assert!(comments[0].contains("- validation_result: `passed`"));
+	assert!(comments[0].contains("- worktree_path: `.worktrees/PUB-618`"));
+}
+
+#[test]
+fn review_handoff_apply_persists_runtime_handoff_marker() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let tracker = FakeTracker::new();
+	let issue = tests::sample_issue();
+	let workflow = tests::sample_workflow();
+	let inspector = FakePullRequestInspector::new(vec![
+		Ok(PullRequestDetails {
+			head_ref_name: String::from("x/decodex-pub-618"),
+			head_ref_oid: String::from("08a20f7dfb9526e7421a5f095b1c6adec84e52d6"),
+			head_repository_name: String::from("decodex"),
+			head_repository_owner: String::from("hack-ink"),
+			is_draft: false,
+			state: String::from("OPEN"),
+			base_ref_name: String::from("main"),
+			url: String::from("https://github.com/hack-ink/decodex/pull/142"),
+		}),
+		Ok(PullRequestDetails {
+			head_ref_name: String::from("x/decodex-pub-618"),
+			head_ref_oid: String::from("08a20f7dfb9526e7421a5f095b1c6adec84e52d6"),
+			head_repository_name: String::from("decodex"),
+			head_repository_owner: String::from("hack-ink"),
+			is_draft: false,
+			state: String::from("OPEN"),
+			base_ref_name: String::from("main"),
+			url: String::from("https://github.com/hack-ink/decodex/pull/142"),
+		}),
+	]);
+	let local_repo_inspector = FakeLocalRepoInspector::new(vec![
+		Ok(tests::sample_local_repo()),
+		Ok(tests::sample_local_repo()),
+	]);
+	let review_context = tests::sample_review_context_in(temp_dir.path());
+	let bridge = TrackerToolBridge::with_review_handoff_for_test(
+		&tracker,
+		&issue,
+		&workflow,
+		review_context.clone(),
+		&inspector,
+		&local_repo_inspector,
+	);
+
+	tests::write_clean_review_checkpoint(&bridge, &issue, &review_context);
+
+	let response = DynamicToolHandler::handle_call(
+		&bridge,
+		ISSUE_REVIEW_HANDOFF_TOOL_NAME,
+		serde_json::json!({
+			"pr_url": "https://github.com/hack-ink/decodex/pull/142",
+			"summary": "Ready for review."
+		}),
+	);
+
+	assert!(response.success);
+
+	bridge.apply_review_handoff().expect("review handoff should apply");
+
+	let marker = tests::persisted_review_handoff_marker(&bridge, &issue, &review_context);
+
+	assert_eq!(marker.branch_name(), review_context.branch_name);
+	assert_eq!(marker.pr_url(), "https://github.com/hack-ink/decodex/pull/142");
+	assert_eq!(marker.pr_head_oid(), "08a20f7dfb9526e7421a5f095b1c6adec84e52d6");
+}
