+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/ghost_lane/mcp_fixture/blocker_protection.rs
@@ -1,0 +1,81 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, FakeTracker, ReviewHandoffMarker, StateStore, TEST_SERVICE_ID, orchestrator,
+};
+
+#[test]
+fn live_operator_status_blocks_missing_issue_ghost_cleanup_when_review_lifecycle_exists() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let marker = ReviewHandoffMarker::new(
+		"run-12",
+		1,
+		"x/pubfi-pub-012",
+		"https://github.com/hack-ink/decodex/pull/12",
+		"main",
+		"x/pubfi-pub-012",
+		"08a20f7dfb9526e7421a5f095b1c6adec84e52d6",
+	);
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_review_handoff_marker(TEST_SERVICE_ID, "PUB-012", &marker)
+		.expect("review lifecycle should record");
+
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should be visible");
+
+	assert_eq!(run.ownership_state, "retained_attention");
+	assert_eq!(run.policy_state, "runtime_recovery_blocked");
+	assert!(run.lane_control_conditions.contains(&String::from("review_lifecycle_present")));
+}
+
+#[test]
+fn live_operator_status_blocks_missing_issue_ghost_cleanup_when_private_evidence_exists() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.append_private_execution_event(
+			"pubfi",
+			"PUB-012",
+			"run-12",
+			1,
+			"diagnostic",
+			serde_json::json!({"schema": "test.private/1"}),
+		)
+		.expect("private evidence should record");
+
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should be visible");
+
+	assert_eq!(run.ownership_state, "retained_attention");
+	assert_eq!(run.policy_state, "runtime_recovery_blocked");
+	assert!(run.lane_control_conditions.contains(&String::from("private_evidence_present")));
+}
