+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/lifecycle/startup_current/child_activity.rs
@@ -1,0 +1,83 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, ChildAgentActivitySummary, OffsetDateTime, ProtocolActivityMarker, StateStore,
+	orchestrator, state,
+};
+
+#[test]
+fn operator_status_snapshot_rolls_current_child_bucket_elapsed_time_into_bucket() {
+	let (_temp_dir, config, _workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = running_lanes::sample_issue("Todo", &[]);
+	let worktree_path = config.worktree_root().join("PUB-101");
+	let started_at = OffsetDateTime::now_utc().unix_timestamp() - 90;
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", &issue.id, "run-1", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			&issue.id,
+			"x/pubfi-pub-101",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-1",
+			attempt_number: 1,
+			thread_id: Some("thread-1"),
+			turn_id: Some("turn-1"),
+			event_count: 1,
+			last_event_type: "item/tool/call",
+			child_agent_activity: Some(&ChildAgentActivitySummary {
+				buckets: vec![state::ChildAgentActivityBucket {
+					name: String::from("Tracker"),
+					event_count: 1,
+					tool_call_count: 1,
+					..state::ChildAgentActivityBucket::default()
+				}],
+				current_bucket: Some(String::from("Tracker")),
+				current_detail: Some(String::from("issue_progress_checkpoint")),
+				current_started_unix_epoch: Some(started_at),
+				current_elapsed_seconds: Some(0),
+				event_count: 1,
+				tool_call_count: 1,
+				..ChildAgentActivitySummary::default()
+			}),
+			protocol_activity: None,
+		},
+	)
+	.expect("protocol activity marker should write");
+
+	let snapshot = orchestrator::build_operator_status_snapshot(&config, &state_store, 10)
+		.expect("snapshot should build");
+	let run = snapshot.current_lanes.first().expect("current lane should exist");
+	let activity = run.child_agent_activity.as_ref().expect("activity should render");
+	let protocol_activity =
+		run.protocol_activity.as_ref().expect("protocol fallback should render");
+	let tracker_bucket =
+		activity.buckets.iter().find(|bucket| bucket.name == "Tracker").expect("tracker bucket");
+
+	assert_eq!(run.wait_reason.as_deref(), Some("tool_execution"));
+	assert_eq!(protocol_activity.waiting_reason.as_deref(), Some("tool_execution"));
+	assert_eq!(
+		snapshot.projects[0].waiting_lane_count, 0,
+		"normal active tool execution is running work, not project-level waiting"
+	);
+	assert_eq!(run.lifecycle_metrics.attempt_count, 1);
+	assert_eq!(run.lifecycle_metrics.captured_attempt_count, 1);
+	assert_eq!(run.lifecycle_metrics.tool_call_count, 1);
+	assert_eq!(run.lifecycle_metrics.phases.len(), 1);
+	assert_eq!(run.lifecycle_metrics.phases[0].phase, "development");
+	assert!(activity.current_elapsed_seconds.is_some_and(|elapsed| elapsed >= 90));
+	assert!(
+		tracker_bucket.wall_seconds >= 90,
+		"current tool-call elapsed time should contribute to tracker bucket wall time"
+	);
+}
